requirements.txt
allure-pytest==2.8.10
allure-python-commons==2.8.10
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
execnet==1.7.1
flaky==3.6.1
glob2==0.7
idna==2.8
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
parse==1.14.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-bdd==3.2.1
pytest-flakefinder==0.1.3
pytest-forked==1.1.3
pytest-xdist==1.31.0
requests==2.22.0
selenium==3.141.0
six==1.14.0
tblib==1.6.0
urllib3==1.25.8
wcwidth==0.1.8
